Issue #1261 Refactor ValidateOpenshiftMinVersion
Rename ValidateOpenshiftMinVersion to IsGTEBaseVersion and
moving it to openshift_versions.go
