+++ pkg/minishift/openshift/version/openshift_versions_test.go
@@ -17,6 +17,7 @@ limitations under the License.
 package version
 
 import (
+	"errors"
 	"io/ioutil"
 	"os"
 	"strings"
@@ -41,7 +42,7 @@ func TestPrintUpStreamVersions(t *testing.T) {
 
 	os.Stdout = f
 	defaultVersion := version.GetOpenShiftVersion()
-	err = PrintUpStreamVersions(f, constants.MinOpenshiftSupportedVersion, defaultVersion)
+	err = PrintUpStreamVersions(f, constants.MinimumSupportedOpenShiftVersion, defaultVersion)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -55,8 +56,8 @@ func TestPrintUpStreamVersions(t *testing.T) {
 		t.Fatal("Error reading file", err)
 	}
 	actualStdout := string(data)
-	if !strings.Contains(actualStdout, constants.MinOpenshiftSupportedVersion) {
-		t.Fatalf("Should Contain constants.MinOpenshiftSupportedVersion in\n %s", actualStdout)
+	if !strings.Contains(actualStdout, constants.MinimumSupportedOpenShiftVersion) {
+		t.Fatalf("Should Contain constants.MinimumSupportedOpenShiftVersion in\n %s", actualStdout)
 	}
 	if !strings.Contains(actualStdout, defaultVersion) {
 		t.Fatalf("Should Contain defaultVersion in\n %s", actualStdout)
@@ -93,5 +94,42 @@ func TestPrintDownStreamVersions(t *testing.T) {
 	actualStdout := string(data)
 	if !strings.Contains(actualStdout, "v3.4.1.10") {
 		t.Fatalf("Should Contain v3.4.1.10 in\n %s", actualStdout)
+	}
+}
+
+func Test_validate_openshift_min_versions(t *testing.T) {
+	var versionTests = []struct {
+		version string // input
+		valid   bool   // expected result
+		err     error
+	}{
+		{"v1.1.0", false, nil},
+		{"v1.2.2", false, nil},
+		{"v1.2.3-beta", false, nil},
+		{"v1.3.1", false, nil},
+		{"v1.3.5-alpha", false, nil},
+		{"foo", false, errors.New("Invalid version format 'foo': No Major.Minor.Patch elements found")},
+		{"151", false, errors.New("Invalid version format '151': No Major.Minor.Patch elements found")},
+		{"v1.4.1", true, nil},
+		{"v1.5.0-alpha.0", true, nil},
+		{"v1.5.1-beta.0", true, nil},
+		{"v3.6.0", true, nil},
+		{"3.6.0", true, nil},
+	}
+
+	minVer := "v1.4.1"
+	for _, versionTest := range versionTests {
+		valid, err := IsGreaterOrEqualToBaseVersion(versionTest.version, minVer)
+		if versionTest.err == nil && err != nil {
+			t.Fatalf("No error expected. Got '%v'", err)
+		}
+
+		if err != nil && err.Error() != versionTest.err.Error() {
+			t.Fatalf("Unexpected error. Expected '%v', got '%v'", versionTest.err, err)
+		}
+
+		if valid != versionTest.valid {
+			t.Fatalf("Expected '%t' Got '%t' for %s", versionTest.valid, valid, versionTest.version)
+		}
 	}
 }
